scripts/main.py
# main.py

import os
import json
import numpy as np
import pandas as pd
import tensorflow as tf
import matplotlib.pyplot as plt
from sklearn.metrics import jaccard_score, confusion_matrix
from ray import tune
import spark

# Configuration
DATA_DIR = 'data/raw/'
PROCESSED_DIR = 'data/processed/'
FIGURES_DIR = 'figures/'
OUTPUTS_DIR = 'outputs/'

# Hyperparameters
hyperparams = {
    'batch_size': 32,
    'epochs': 50,
    'learning_rate': 0.001,
    'dropout_rate': 0.5,
    'num_filters': [32, 64, 128],
    'kernel_size': 3,
    'pool_size': 2,
    'dense_units': 128
}

# Function Placeholders

def load_data():
    pass

def preprocess_data():
    pass

def build_model(hp):
    pass

def train_model(model, train_data, val_data, hp):
    pass

def evaluate_model(model, test_data):
    pass

def plot_confusion_matrix(cm, classes, filename):
    pass

def plot_jaccard_score(jaccard_scores, filename):
    pass

def plot_loss_vs_epochs(history, filename):
    pass

def tune_hyperparameters(config):
    pass

def save_hyperparameters_performance(hyperparams, performance, filename):
    pass

def main():
    # Load Data
    train_data, val_data, test_data = load_data()

    # Preprocess Data
    train_data, val_data, test_data = preprocess_data(train_data, val_data, test_data)

    # Tune Hyperparameters
    analysis = tune.run(
        tune.with_parameters(train_model, train_data=train_data, val_data=val_data),
        config=hyperparams,
        resources_per_trial={'cpu': 2, 'gpu': 0},
        metric="mean_accuracy",
        mode="max"
    )

    # Get best hyperparameters
    best_hyperparams = analysis.get_best_config(metric="mean_accuracy", mode="max")

    # Save best hyperparameters
    save_hyperparameters_performance(best_hyperparams, analysis, os.path.join(OUTPUTS_DIR, 'best_hyperparams.json'))

    # Build Model with best hyperparameters
    model = build_model(best_hyperparams)

    # Train Model with best hyperparameters
    history = train_model(model, train_data, val_data, best_hyperparams)

    # Evaluate Model
    performance = evaluate_model(model, test_data)

    # Plot Figures
    cm = confusion_matrix(test_data.labels, model.predict(test_data.images))
    plot_confusion_matrix(cm, classes=train_data.classes, filename=os.path.join(FIGURES_DIR, 'confusion_matrix.png'))
    plot_jaccard_score(performance['jaccard_scores'], filename=os.path.join(FIGURES_DIR, 'jaccard_scores.png'))
    plot_loss_vs_epochs(history, filename=os.path.join(FIGURES_DIR, 'loss_vs_epochs.png'))

if __name__ == "__main__":
    main()
