requirements.txt
tensorflow
numpy
pandas
scikit-learn
matplotlib
ray[tune]
json
apache-spark